Moved the mongodb file store to the integrations submodule.
