+++ integrations/datastore/pom.xml
@@ -1,4 +1,4 @@
-<?xml version="1.0" encoding="UTF-8"?>
+<?xml version="1.0"?>
 <!--
   ~ Copyright (c) 2021 Kaiserpfalz EDV-Service, Roland T. Lichti.
   ~
@@ -16,40 +16,40 @@
   ~ along with this program.  If not, see <https://www.gnu.org/licenses/>.
   -->
 
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
     <parent>
-        <artifactId>kp-rpg</artifactId>
+        <artifactId>integrations</artifactId>
         <groupId>de.kaiserpfalz-edv.rpg</groupId>
         <version>1.2.0-SNAPSHOT</version>
         <relativePath>../pom.xml</relativePath>
     </parent>
-    <artifactId>rpg-core</artifactId>
-    <name>KP RPG :: Core Services</name>
-    <description>The services used by almost all role playing systems.</description>
+    <artifactId>datastore</artifactId>
+    <name>KP RPG :: Integration :: DataStore</name>
+    <description>The data store for TOMB.</description>
 
+    <properties>
+        <activate.jcenter>true</activate.jcenter>
+    </properties>
 
     <dependencies>
         <dependency>
-            <groupId>net.objecthunter</groupId>
-            <artifactId>exp4j</artifactId>
+            <groupId>de.kaiserpfalz-edv.rpg</groupId>
+            <artifactId>rpg-core</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>io.quarkus</groupId>
+            <artifactId>quarkus-mongodb-client</artifactId>
         </dependency>
         <dependency>
             <groupId>io.quarkus</groupId>
-            <artifactId>quarkus-resteasy-jackson</artifactId>
+            <artifactId>quarkus-mongodb-panache</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>de.kaiserpfalz-edv.rpg.test-support</groupId>
+            <artifactId>mongodb-testsupport</artifactId>
         </dependency>
     </dependencies>
-
-    <dependencyManagement>
-        <dependencies>
-            <dependency>
-                <groupId>org.testcontainers</groupId>
-                <artifactId>testcontainers-bom</artifactId>
-                <version>1.15.1</version>
-                <type>pom</type>
-                <scope>import</scope>
-            </dependency>
-        </dependencies>
-    </dependencyManagement>
 </project>
