+++ integrations/datastore/src/main/java/de/kaiserpfalzedv/rpg/integrations/datastore/FileDateabase.java
@@ -1,0 +1,52 @@
+/*
+ * Copyright (c) 2021 Kaiserpfalz EDV-Service, Roland T. Lichti.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package de.kaiserpfalzedv.rpg.integrations.datastore;
+
+import com.mongodb.client.MongoClient;
+import com.mongodb.client.gridfs.GridFSBucket;
+import com.mongodb.client.gridfs.GridFSBuckets;
+
+import javax.enterprise.context.ApplicationScoped;
+import javax.enterprise.inject.Produces;
+import javax.inject.Inject;
+
+/**
+ * The provider for the GrifFS to save files to.
+ *
+ * @author klenkes74 {@literal <rlichti@kaiserpfalz-edv.de>}
+ * @since 1.0.0 2021-01-08
+ */
+@ApplicationScoped
+public class FileDateabase {
+    private static final String FILE_DATABASE_NAME = "files";
+
+    /**
+     * The mongo db client used for creating the GridFS bucket.
+     */
+    @SuppressWarnings("CdiInjectionPointsInspection")
+    @Inject
+    MongoClient client;
+
+    /**
+     * @return The bucket to handle the files of the service.
+     */
+    @Produces
+    public GridFSBucket fileBucket() {
+        return GridFSBuckets.create(client.getDatabase(FILE_DATABASE_NAME));
+    }
+}
