+++ integrations/datastore/src/main/java/de/kaiserpfalzedv/rpg/integrations/datastore/MongoDBFileResources.java
@@ -15,14 +15,14 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package de.kaiserpfalzedv.rpg.core.files.store;
+package de.kaiserpfalzedv.rpg.integrations.datastore;
 
 import com.mongodb.MongoGridFSException;
 import com.mongodb.client.gridfs.GridFSBucket;
-import de.kaiserpfalzedv.rpg.core.files.FileCouldNotBeDeletedException;
-import de.kaiserpfalzedv.rpg.core.files.FileCouldNotBeSavedException;
-import de.kaiserpfalzedv.rpg.core.files.FileNotFoundException;
-import de.kaiserpfalzedv.rpg.core.files.FileResourceService;
+import de.kaiserpfalzedv.rpg.integrations.datastore.store.FileCouldNotBeDeletedException;
+import de.kaiserpfalzedv.rpg.integrations.datastore.store.FileCouldNotBeSavedException;
+import de.kaiserpfalzedv.rpg.integrations.datastore.store.FileNotFoundException;
+import de.kaiserpfalzedv.rpg.integrations.datastore.store.FileResourceService;
 import org.bson.BsonString;
 import org.bson.BsonValue;
