+++ integrations/datastore/src/test/resources/application.properties
@@ -37,6 +37,11 @@ quarkus.swagger-ui.path=/docs
 quarkus.swagger-ui.always-include=true
 quarkus.swagger-ui.enable=true
 #
+quarkus.mongodb.connection-string=tomb-mongodb
+quarkus.mongodb.credentials.username=<username>
+quarkus.mongodb.credentials.password=<password>
+quarkus.mongodb.database=tomb
+#
 quarkus.oauth2.enabled=false
 quarkus.oauth2.role-claim=bot,connections,email,guilds,guilds.join,messages.read,applications.entitlements,applications.commands.update,rpc,rpc.api,rpc.notifications.read
 quarkus.oauth2.client-id=<client-id>
